requirements.txt
flask>=3.0.0
pyzbar>=0.1.9
Pillow>=10.0.0
gunicorn>=21.2.0
